+++ src/main/java/com/pricing/domain/exception/ExceptionType.java
@@ -4,6 +4,7 @@ public enum ExceptionType {
 
   REPOSITORY,
   RULES_VIOLATION,
+  NOT_FOUND,
   UNKNOWN
 
 }